requirements.txt
pandas
requests
psycopg2-binary
python-dotenv
transformers
torch
scikit-learn
newspaper3k
beautifulsoup4
sentence-transformers
keybert
google-generativeai
streamlit
bert-extractive-summarizer
spacy
sumy
nltk